refactor:api폴더 구조 변경,type 구조 변경
