+++ src/api/activities/index.ts
@@ -1,0 +1,7 @@
+import { instance } from "@/lib/axios";
+import { Activity } from "@/types/activities";
+
+export const getAcitivity = async (id: any): Promise<Activity> => {
+  const res = await instance.get(`/activities/${id}`);
+  return res.data;
+};
