+++ src/api/myActivities/index.ts
@@ -1,0 +1,5 @@
+import { instance } from "@/lib/axios";
+
+export const deleteActivity = async ({ id }: any) => {
+  return await instance.delete(`/my-activities/${id}`);
+};
