+++ src/hooks/activities/index.ts
@@ -1,0 +1,8 @@
+import { getAcitivity } from "@/api/activities";
+import { queryKey } from "@/constants/queryKeys";
+import { useQuery } from "@tanstack/react-query";
+
+export const useGetActivity = useQuery({
+  queryKey: [queryKey.activity],
+  queryFn: (id) => getAcitivity(id)
+});
