+++ src/hooks/users/index.ts
@@ -1,0 +1,8 @@
+import { getUser } from "@/api/users";
+import { queryKey } from "@/constants/queryKeys";
+import { useQuery } from "@tanstack/react-query";
+
+export const useGetUser = useQuery({
+  queryKey: [queryKey.usersMe],
+  queryFn: () => getUser()
+});
