+++ src/types/common/index.ts
@@ -1,0 +1,8 @@
+export type Category =
+  | "문화 · 예술"
+  | "식음료"
+  | "스포츠"
+  | "투어"
+  | "관광"
+  | "웰빙"
+  | "";
